12.function/1.new 하위 수업 추가
