+++ 4.study/11.function/1.new/6.Login_Ruby/1.rb
@@ -1,16 +1,8 @@
 # 로그인 애플리케이션(루비)
+require_relative 'Auth'
 puts("id를 입력해 주세요.")
 input_id = gets.chomp()
-def login(id)
-    members = ['kst', 'gst', 'kim']
-    for member in members do
-        if member == id
-            return true
-        end
-    end
-    return false
-end
-if login(input_id)
+if Auth.login(input_id)
     puts('Hello, '+input_id)
 else
     puts('Who are you?')
