+++ 4.study/12.module/1.new/6.Login_application/auth.py
@@ -1,0 +1,6 @@
+import auth
+input_id = input("id를 입력해 주세요.\n")
+def auth.login(input_id):
+    print('Hello, '+input_id)
+else:
+    print('Who are you?')
