+++ 4.study/12.module/1.new/6.Login_application/auth.rb
@@ -1,0 +1,12 @@
+module Auth
+    module_function()
+def login(id)
+    members = ['kst', 'gst', 'kim']
+    for member in members do
+        if member == id
+            return true
+        end
+            end
+        return false
+    end
+end
